refactor(main): remove bundles mod
- Also give more flexibility with algorithms' steps

+++ src/ui/constants/mod.rs
@@ -1,3 +1,3 @@
-pub const FLOOR: u32 = 1;
-pub const CEIL: u32 = 11;
+pub const FLOOR: usize = 1;
+pub const CEIL: usize = 11;
 pub const VECTOR_SIZE: usize = 5;
